more longs
git-svn-id: http://pinocchio.unibe.ch/svn/pinocchio@1538 475e7f6d-5e0e-419c-b4a3-23a65b5ccdbc

+++ src/system/number/Float.c
@@ -69,7 +69,7 @@ NATIVE0(Float_hash)
 String Float_asString(float self, uns_int base)
 {
     char *chrs;
-    int size = asprintf(&chrs, "%f", self);
+    long size = asprintf(&chrs, "%f", self);
     assert1(size != -1, "Unable to convert float to string");
     String result = new_String(ascii_to_unicode(chrs));
     free(chrs);
